+++ testcases/integration-tests/check_untagged.sh
@@ -28,7 +28,12 @@ prepare_check_untagged()
 call_check_untagged()
 {
 	local check_dir="$1"
-	kphotoalbum --config "$check_dir/kphotoalbumrc" --db "$check_dir/db/index.xml" --search
+	local automatic
+	if [ -n "$NON_INTERACTIVE" ]
+	then
+		automatic="--save-and-quit"
+	fi
+	kphotoalbum $automatic --config "$check_dir/kphotoalbumrc" --db "$check_dir/db/index.xml" --search
 }
 check_untagged()
 {